rename util for getting device measurements and discard unused functions
